Deep-Learning-Image-Style-Transfer/requirements.txt
numpy==1.21.6
Flask==3.0.3
gunicorn==22.0.0
torch==2.2.0
torchvision==0.17.0
Pillow==10.2.0
